+++ src/serviceUnavailableInterceptor/serviceUnavailableInterceptorOptions.ts
@@ -7,7 +7,7 @@ import {ERROR_HANDLING_NOTIFICATIONS} from '../misc/tokens';
 /**
  * Action called when 503 http status code is intercepted
  */
-export type ServiceUnavailableAction = (injector: Injector, observer: Observer<void>) => void;
+export type ServiceUnavailableAction = (injector: Injector, observer: Observer<any>) => void;
 
 /**
  * Options for ServiceUnavailableInterceptorOptions
@@ -17,12 +17,12 @@ export class ServiceUnavailableInterceptorOptions
     /**
      * Text displayed
      */
-    text?: string = 'Remote server is unavailable. Try again later.';
+    text: string = 'Remote server is unavailable. Try again later.';
 
     /**
      * Action called when 504 status code received
      */
-    action?: ServiceUnavailableAction = (injector: Injector, observer: Observer<any>) =>
+    action: ServiceUnavailableAction = (injector: Injector, observer: Observer<any>) =>
     {
         const notifications = injector.get(ERROR_HANDLING_NOTIFICATIONS);